Online/functions/requirements.txt
azure-functions
azure-cognitiveservices-vision-computervision